% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ordered}
\alias{ordered}
\title{A order-located function}
\usage{
ordered(vec)
}
\arguments{
\item{vec}{a numeric vector}
}
\value{
an integer vector with i-th element equaling the index (starting with 0) of the i-the smallest element in the input vector
}
\description{
A function returns the index (starting with 0) of the i-th smallest element in the vector in its i-th element
}
\examples{
\dontrun{
x<-c(1,3,2,6)
ordered(x)}
}
