\name{StatComp22003-package}
\alias{StatComp22003-package}
\alias{StatComp22003}
\docType{package}
\title{
  A random split tree
}
\description{
  StatComp22003 is a simple R package developed to build a tree for binary classification problem for the 'Statistical Computing' course. 
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Jiaqiang Li.

Maintainer: Jiaqiang Li <ljq1492@mail.ustc.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }


