% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tree_predict}
\alias{tree_predict}
\title{Predict the label of a sample using our tree}
\usage{
tree_predict(node_info, x_test)
}
\arguments{
\item{node_info}{A tree producted by \code{tree_create}}

\item{x_test}{A vector, the sample}
}
\value{
The predict label of x_test
}
\description{
Predict the label of a sample using our tree
}
\examples{
\dontrun{
data(iris)
iris_data <- as.matrix(iris[1:100,-5])
iris_label <- as.numeric(iris[1:100,5])-1
set.seed(15)
ind <- 1:100
ind_tr <- sample(ind,60)
ind_test <- ind[-ind_tr]
t <- tree_create(x=iris_data,y=iris_label,q=4,random_split=F)
pre <- sapply(1:40,tree_predict(t,iris_data[ind_test[i],]))
}
}
